+++ CHANGELOG.rst
@@ -31,6 +31,7 @@ New features and notable changes:
   more than once. (:issue:`1129`)
 - Add :option:`--sonarqube-metric` to define which metric is used as branch metric in report. (:issue:`1132`)
 - Add :option:`--sonarqube-pretty` to pretty print the report. (:issue:`1136`)
+- If the value for :option:`-j` is 0, the number of CPUs is used. If it is negative the it means "all but N CPUs". (:issue:`1148`)
 
 Bug fixes and small improvements: